[IN-18] fix forgot-password page component
