+++ src/components/forgot-password/forgot-password.module.scss
@@ -1,0 +1,40 @@
+div.forgotPassword {
+  display: flex;
+  flex-direction: column;
+  flex-shrink: 0;
+  align-items: center;
+
+  width: 378px;
+  height: 456px;
+  margin: 72px auto 0;
+  padding: 24px;
+
+  background: var(--dark-500);
+  border: 1px solid var(--dark-300);
+  border-radius: 2px;
+
+  .form {
+    width: 100%;
+  }
+
+  .title {
+    margin-bottom: 37px;
+  }
+
+  .instructions {
+    margin-bottom: 17px;
+    color: var(--light-900);
+  }
+
+  .submitBtn,
+  .returnBtn {
+    margin-bottom: 24px;
+  }
+
+  .returnBtn {
+    display: block;
+    width: fit-content;
+    margin-right: auto;
+    margin-left: auto;
+  }
+}
