+++ src/pages/forgot-password/index.tsx
@@ -1,14 +1,14 @@
 import React from 'react'
 
+import ForgotPasswordPageComponent from '@/components/forgot-password/forgot-password'
 import { HeadMeta } from '@/components/head-meta'
-import ForgotPasswordPageComponent from '@/page-components/forgot-password/forgot-password'
 
 const ForgotPassword = () => {
   return (
     <>
       <HeadMeta title="Forgot password" />
       <main>
-        <ForgotPasswordPageComponent></ForgotPasswordPageComponent>
+        <ForgotPasswordPageComponent />
       </main>
     </>
   )
